Removed a few explicit casts which caused warnings. (From louisdx)
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@3707 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/libs/heap/heap.c
@@ -49,7 +49,7 @@ Heap_new(HeapValue_Comparator comparator, size_t initialSize, size_t minSize,
 	heap->minSize = minSize;
 	heap->minFillQuotient = minFillQuotient;
 	heap->size = nextPower2(initialSize);
-	heap->minFill = (size_t) ceil(((double) (heap->size >> 1))
+	heap->minFill = ceil(((double) (heap->size >> 1))
 			* heap->minFillQuotient);
 	heap->entries = malloc(heap->size * sizeof (HeapValue *));
 	heap->numEntries = 0;
